fix: Endpoints de posts e comments agora permitem apenas que o seu dono os modifique ou apague

+++ src/use-cases/comments/delete-comment.ts
@@ -1,21 +1,32 @@
 import type { CommentsRepository } from "@/repositories/comments-repository.js"
+import type { UsersRepository } from "@/repositories/users-repository.js"
+import { NotAllowedError } from "../errors/not-allowed-error.js"
 import { ResourceNotFoundError } from "../errors/resource-not-found-error.js"
 
 interface DeleteCommentUseCaseRequest {
     publicID: string
+    requesterPublicId: string
 }
 
 export class DeleteCommentUseCase {
-    constructor(private commentsRepository: CommentsRepository) {}
+    constructor(
+        private commentsRepository: CommentsRepository,
+        private usersRepository: UsersRepository,
+    ) {}
     async execute({
-        publicID
+        publicID,
+        requesterPublicId,
     }: DeleteCommentUseCaseRequest) {
-        const comment = await this.commentsRepository.findBy({
-            publicID,
-        })
+        const comment = await this.commentsRepository.findBy({ publicID })
 
         if (!comment) {
             throw new ResourceNotFoundError()
+        }
+
+        const requester = await this.usersRepository.findBy({ publicID: requesterPublicId })
+
+        if (requester?.id !== comment.authorId) {
+            throw new NotAllowedError()
         }
 
         await this.commentsRepository.delete(comment.id)
